// adapters/edc/overlays/dashboard/projects/dashboard-core/assets/src/models/ontology.ts
export interface Ontology {
  uri: string;
  nsp: string;
  prefix: string;
  titles: OntologyTitle[];
  versions: OntologyVersion[];
  artifacts: OntologyArtifacts;
}

export interface OntologyTitle {
  value: string;
  lang: string;
  _id: string;
}

export interface OntologyVersion {
  name: string;
  issued: string;
  isReviewed: boolean;
}

export interface OntologyArtifacts {
  shapes: string[];
}

export const NEW_SHACLE_FILE_VALUE = 'newShacleFile';

export interface OntologyAssetSelection {
  ontologyUri: string;
  ontologyVersion: string;
  ontologyShacl: string;
  ontologyDownloadUrl: string;
  shaclDownloadUrl: string;
  isOntologyComplete: boolean;
}